Fix various compilation and linker issues across several platforms

+++ panda/src/pgraph/renderState.cxx
@@ -985,7 +985,7 @@ clear_cache() {
       RenderState *state = (RenderState *)(*ti).p();
 
       int i;
-      int cache_size = state->_composition_cache.get_size();
+      int cache_size = (int)state->_composition_cache.get_size();
       for (i = 0; i < cache_size; ++i) {
         if (state->_composition_cache.has_element(i)) {
           const RenderState *result = state->_composition_cache.get_data(i)._result;
@@ -995,10 +995,10 @@ clear_cache() {
           }
         }
       }
-      _cache_stats.add_total_size(-state->_composition_cache.get_num_entries());
+      _cache_stats.add_total_size(-(int)state->_composition_cache.get_num_entries());
       state->_composition_cache.clear();
 
-      cache_size = state->_invert_composition_cache.get_size();
+      cache_size = (int)state->_invert_composition_cache.get_size();
       for (i = 0; i < cache_size; ++i) {
         if (state->_invert_composition_cache.has_element(i)) {
           const RenderState *result = state->_invert_composition_cache.get_data(i)._result;
@@ -1008,7 +1008,7 @@ clear_cache() {
           }
         }
       }
-      _cache_stats.add_total_size(-state->_invert_composition_cache.get_num_entries());
+      _cache_stats.add_total_size(-(int)state->_invert_composition_cache.get_num_entries());
       state->_invert_composition_cache.clear();
     }
