+++ panda/src/putil/bitArray.cxx
@@ -575,8 +575,8 @@ output_binary(ostream &out, int spaces_every) const {
   if (_highest_bits) {
     out << "...1 ";
   }
-  size_t num_bits = max(get_num_bits(), (size_t)spaces_every);
-  for (size_t i = num_bits - 1; i >= 0; i--) {
+  int num_bits = max((int)get_num_bits(), spaces_every);
+  for (int i = num_bits - 1; i >= 0; i--) {
     if (spaces_every != 0 && ((i % spaces_every) == spaces_every - 1)) {
       out << ' ';
     }
